+++ gems/aws-sdk-directoryservice/lib/aws-sdk-directoryservice/client_api.rb
@@ -89,6 +89,7 @@ module Aws::DirectoryService
     CreatedDateTime = Shapes::TimestampShape.new(name: 'CreatedDateTime')
     CustomerId = Shapes::StringShape.new(name: 'CustomerId')
     CustomerUserName = Shapes::StringShape.new(name: 'CustomerUserName')
+    DataAccessStatus = Shapes::StringShape.new(name: 'DataAccessStatus')
     DeleteAssociatedConditionalForwarder = Shapes::BooleanShape.new(name: 'DeleteAssociatedConditionalForwarder')
     DeleteConditionalForwarderRequest = Shapes::StructureShape.new(name: 'DeleteConditionalForwarderRequest')
     DeleteConditionalForwarderResult = Shapes::StructureShape.new(name: 'DeleteConditionalForwarderResult')
@@ -112,6 +113,8 @@ module Aws::DirectoryService
     DescribeConditionalForwardersResult = Shapes::StructureShape.new(name: 'DescribeConditionalForwardersResult')
     DescribeDirectoriesRequest = Shapes::StructureShape.new(name: 'DescribeDirectoriesRequest')
     DescribeDirectoriesResult = Shapes::StructureShape.new(name: 'DescribeDirectoriesResult')
+    DescribeDirectoryDataAccessRequest = Shapes::StructureShape.new(name: 'DescribeDirectoryDataAccessRequest')
+    DescribeDirectoryDataAccessResult = Shapes::StructureShape.new(name: 'DescribeDirectoryDataAccessResult')
     DescribeDomainControllersRequest = Shapes::StructureShape.new(name: 'DescribeDomainControllersRequest')
     DescribeDomainControllersResult = Shapes::StructureShape.new(name: 'DescribeDomainControllersResult')
     DescribeEventTopicsRequest = Shapes::StructureShape.new(name: 'DescribeEventTopicsRequest')
@@ -166,6 +169,8 @@ module Aws::DirectoryService
     DirectoryVpcSettingsDescription = Shapes::StructureShape.new(name: 'DirectoryVpcSettingsDescription')
     DisableClientAuthenticationRequest = Shapes::StructureShape.new(name: 'DisableClientAuthenticationRequest')
     DisableClientAuthenticationResult = Shapes::StructureShape.new(name: 'DisableClientAuthenticationResult')
+    DisableDirectoryDataAccessRequest = Shapes::StructureShape.new(name: 'DisableDirectoryDataAccessRequest')
+    DisableDirectoryDataAccessResult = Shapes::StructureShape.new(name: 'DisableDirectoryDataAccessResult')
     DisableLDAPSRequest = Shapes::StructureShape.new(name: 'DisableLDAPSRequest')
     DisableLDAPSResult = Shapes::StructureShape.new(name: 'DisableLDAPSResult')
     DisableRadiusRequest = Shapes::StructureShape.new(name: 'DisableRadiusRequest')
@@ -182,6 +187,8 @@ module Aws::DirectoryService
     DomainControllers = Shapes::ListShape.new(name: 'DomainControllers')
     EnableClientAuthenticationRequest = Shapes::StructureShape.new(name: 'EnableClientAuthenticationRequest')
     EnableClientAuthenticationResult = Shapes::StructureShape.new(name: 'EnableClientAuthenticationResult')
+    EnableDirectoryDataAccessRequest = Shapes::StructureShape.new(name: 'EnableDirectoryDataAccessRequest')
+    EnableDirectoryDataAccessResult = Shapes::StructureShape.new(name: 'EnableDirectoryDataAccessResult')
     EnableLDAPSRequest = Shapes::StructureShape.new(name: 'EnableLDAPSRequest')
     EnableLDAPSResult = Shapes::StructureShape.new(name: 'EnableLDAPSResult')
     EnableRadiusRequest = Shapes::StructureShape.new(name: 'EnableRadiusRequest')
@@ -655,6 +662,12 @@ module Aws::DirectoryService
     DescribeDirectoriesResult.add_member(:next_token, Shapes::ShapeRef.new(shape: NextToken, location_name: "NextToken"))
     DescribeDirectoriesResult.struct_class = Types::DescribeDirectoriesResult
 
+    DescribeDirectoryDataAccessRequest.add_member(:directory_id, Shapes::ShapeRef.new(shape: DirectoryId, required: true, location_name: "DirectoryId"))
+    DescribeDirectoryDataAccessRequest.struct_class = Types::DescribeDirectoryDataAccessRequest
+
+    DescribeDirectoryDataAccessResult.add_member(:data_access_status, Shapes::ShapeRef.new(shape: DataAccessStatus, location_name: "DataAccessStatus"))
+    DescribeDirectoryDataAccessResult.struct_class = Types::DescribeDirectoryDataAccessResult
+
     DescribeDomainControllersRequest.add_member(:directory_id, Shapes::ShapeRef.new(shape: DirectoryId, required: true, location_name: "DirectoryId"))
     DescribeDomainControllersRequest.add_member(:domain_controller_ids, Shapes::ShapeRef.new(shape: DomainControllerIds, location_name: "DomainControllerIds"))
     DescribeDomainControllersRequest.add_member(:next_token, Shapes::ShapeRef.new(shape: NextToken, location_name: "NextToken"))
@@ -845,6 +858,11 @@ module Aws::DirectoryService
 
     DisableClientAuthenticationResult.struct_class = Types::DisableClientAuthenticationResult
 
+    DisableDirectoryDataAccessRequest.add_member(:directory_id, Shapes::ShapeRef.new(shape: DirectoryId, required: true, location_name: "DirectoryId"))
+    DisableDirectoryDataAccessRequest.struct_class = Types::DisableDirectoryDataAccessRequest
+
+    DisableDirectoryDataAccessResult.struct_class = Types::DisableDirectoryDataAccessResult
+
     DisableLDAPSRequest.add_member(:directory_id, Shapes::ShapeRef.new(shape: DirectoryId, required: true, location_name: "DirectoryId"))
     DisableLDAPSRequest.add_member(:type, Shapes::ShapeRef.new(shape: LDAPSType, required: true, location_name: "Type"))
     DisableLDAPSRequest.struct_class = Types::DisableLDAPSRequest
@@ -890,6 +908,11 @@ module Aws::DirectoryService
     EnableClientAuthenticationRequest.struct_class = Types::EnableClientAuthenticationRequest
 
     EnableClientAuthenticationResult.struct_class = Types::EnableClientAuthenticationResult
+
+    EnableDirectoryDataAccessRequest.add_member(:directory_id, Shapes::ShapeRef.new(shape: DirectoryId, required: true, location_name: "DirectoryId"))
+    EnableDirectoryDataAccessRequest.struct_class = Types::EnableDirectoryDataAccessRequest
+
+    EnableDirectoryDataAccessResult.struct_class = Types::EnableDirectoryDataAccessResult
 
     EnableLDAPSRequest.add_member(:directory_id, Shapes::ShapeRef.new(shape: DirectoryId, required: true, location_name: "DirectoryId"))
     EnableLDAPSRequest.add_member(:type, Shapes::ShapeRef.new(shape: LDAPSType, required: true, location_name: "Type"))
@@ -1744,6 +1767,19 @@ module Aws::DirectoryService
         )
       end)
 
+      api.add_operation(:describe_directory_data_access, Seahorse::Model::Operation.new.tap do |o|
+        o.name = "DescribeDirectoryDataAccess"
+        o.http_method = "POST"
+        o.http_request_uri = "/"
+        o.input = Shapes::ShapeRef.new(shape: DescribeDirectoryDataAccessRequest)
+        o.output = Shapes::ShapeRef.new(shape: DescribeDirectoryDataAccessResult)
+        o.errors << Shapes::ShapeRef.new(shape: DirectoryDoesNotExistException)
+        o.errors << Shapes::ShapeRef.new(shape: UnsupportedOperationException)
+        o.errors << Shapes::ShapeRef.new(shape: AccessDeniedException)
+        o.errors << Shapes::ShapeRef.new(shape: ClientException)
+        o.errors << Shapes::ShapeRef.new(shape: ServiceException)
+      end)
+
       api.add_operation(:describe_domain_controllers, Seahorse::Model::Operation.new.tap do |o|
         o.name = "DescribeDomainControllers"
         o.http_method = "POST"
@@ -1922,6 +1958,21 @@ module Aws::DirectoryService
         o.errors << Shapes::ShapeRef.new(shape: ServiceException)
       end)
 
+      api.add_operation(:disable_directory_data_access, Seahorse::Model::Operation.new.tap do |o|
+        o.name = "DisableDirectoryDataAccess"
+        o.http_method = "POST"
+        o.http_request_uri = "/"
+        o.input = Shapes::ShapeRef.new(shape: DisableDirectoryDataAccessRequest)
+        o.output = Shapes::ShapeRef.new(shape: DisableDirectoryDataAccessResult)
+        o.errors << Shapes::ShapeRef.new(shape: DirectoryDoesNotExistException)
+        o.errors << Shapes::ShapeRef.new(shape: DirectoryUnavailableException)
+        o.errors << Shapes::ShapeRef.new(shape: UnsupportedOperationException)
+        o.errors << Shapes::ShapeRef.new(shape: DirectoryInDesiredStateException)
+        o.errors << Shapes::ShapeRef.new(shape: AccessDeniedException)
+        o.errors << Shapes::ShapeRef.new(shape: ClientException)
+        o.errors << Shapes::ShapeRef.new(shape: ServiceException)
+      end)
+
       api.add_operation(:disable_ldaps, Seahorse::Model::Operation.new.tap do |o|
         o.name = "DisableLDAPS"
         o.http_method = "POST"
@@ -1972,6 +2023,21 @@ module Aws::DirectoryService
         o.errors << Shapes::ShapeRef.new(shape: InvalidClientAuthStatusException)
         o.errors << Shapes::ShapeRef.new(shape: AccessDeniedException)
         o.errors << Shapes::ShapeRef.new(shape: NoAvailableCertificateException)
+        o.errors << Shapes::ShapeRef.new(shape: ClientException)
+        o.errors << Shapes::ShapeRef.new(shape: ServiceException)
+      end)
+
+      api.add_operation(:enable_directory_data_access, Seahorse::Model::Operation.new.tap do |o|
+        o.name = "EnableDirectoryDataAccess"
+        o.http_method = "POST"
+        o.http_request_uri = "/"
+        o.input = Shapes::ShapeRef.new(shape: EnableDirectoryDataAccessRequest)
+        o.output = Shapes::ShapeRef.new(shape: EnableDirectoryDataAccessResult)
+        o.errors << Shapes::ShapeRef.new(shape: DirectoryDoesNotExistException)
+        o.errors << Shapes::ShapeRef.new(shape: DirectoryUnavailableException)
+        o.errors << Shapes::ShapeRef.new(shape: UnsupportedOperationException)
+        o.errors << Shapes::ShapeRef.new(shape: DirectoryInDesiredStateException)
+        o.errors << Shapes::ShapeRef.new(shape: AccessDeniedException)
         o.errors << Shapes::ShapeRef.new(shape: ClientException)
         o.errors << Shapes::ShapeRef.new(shape: ServiceException)
       end)