+++ gems/aws-sdk-directoryservice/lib/aws-sdk-directoryservice/plugins/endpoints.rb
@@ -117,6 +117,8 @@ module Aws::DirectoryService
             Aws::DirectoryService::Endpoints::DescribeConditionalForwarders.build(context)
           when :describe_directories
             Aws::DirectoryService::Endpoints::DescribeDirectories.build(context)
+          when :describe_directory_data_access
+            Aws::DirectoryService::Endpoints::DescribeDirectoryDataAccess.build(context)
           when :describe_domain_controllers
             Aws::DirectoryService::Endpoints::DescribeDomainControllers.build(context)
           when :describe_event_topics
@@ -137,6 +139,8 @@ module Aws::DirectoryService
             Aws::DirectoryService::Endpoints::DescribeUpdateDirectory.build(context)
           when :disable_client_authentication
             Aws::DirectoryService::Endpoints::DisableClientAuthentication.build(context)
+          when :disable_directory_data_access
+            Aws::DirectoryService::Endpoints::DisableDirectoryDataAccess.build(context)
           when :disable_ldaps
             Aws::DirectoryService::Endpoints::DisableLDAPS.build(context)
           when :disable_radius
@@ -145,6 +149,8 @@ module Aws::DirectoryService
             Aws::DirectoryService::Endpoints::DisableSso.build(context)
           when :enable_client_authentication
             Aws::DirectoryService::Endpoints::EnableClientAuthentication.build(context)
+          when :enable_directory_data_access
+            Aws::DirectoryService::Endpoints::EnableDirectoryDataAccess.build(context)
           when :enable_ldaps
             Aws::DirectoryService::Endpoints::EnableLDAPS.build(context)
           when :enable_radius